% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{barplot_bppercent}
\alias{barplot_bppercent}
\title{Bar Plot for Base Pair Percentage}
\usage{
barplot_bppercent(matrix)
}
\arguments{
\item{matrix}{A matrix of bp percentage features (rows represent samples).}
}
\value{
A ggplot2 plot of the bp percentages.
}
\description{
Creates a bar plot of the mean percentage of base pairs for each bp combination,
with individual data points overlaid.
}
