% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{create_exon_df}
\alias{create_exon_df}
\title{Create Exon Feature DataFrame}
\usage{
create_exon_df(seq_exons, upstr_exons, label = NA)
}
\arguments{
\item{seq_exons}{An RNAStringSet object of exon sequences.}

\item{upstr_exons}{A GRanges object representing upstream exon regions.}

\item{label}{(Optional) A binary label for machine learning purposes.}
}
\value{
A data frame with all the computed features.
}
\description{
Combines nucleotide percentage features, di-nucleotide and tri-nucleotide features,
and sliding window features for exons. It also adds an auxiliary ID and an optional label.
}
