% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{get_bp_count}
\alias{get_bp_count}
\title{Get Base Pair Counts}
\usage{
get_bp_count(stringSet)
}
\arguments{
\item{stringSet}{An RNAStringSet object.}
}
\value{
A matrix with counts of each nucleotide.
}
\description{
Computes the count (frequency) of each nucleotide (A, U, C, G) for each sequence
in an RNAStringSet object.
}
