% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{get_bp_percent}
\alias{get_bp_percent}
\title{Get Base Pair Percentages}
\usage{
get_bp_percent(stringSet)
}
\arguments{
\item{stringSet}{An RNAStringSet object.}
}
\value{
A matrix with percentage (proportions) of each nucleotide.
}
\description{
Computes the percentage of each nucleotide (A, U, C, G) in each sequence of an
RNAStringSet object.
}
