% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{get_dibp_features}
\alias{get_dibp_features}
\title{Get Di-nucleotide Features}
\usage{
get_dibp_features(stringSet)
}
\arguments{
\item{stringSet}{An RNAStringSet object.}
}
\value{
A matrix with percentages for each di-nucleotide.
}
\description{
Calculates the frequency of di-nucleotides (pairs, e.g. AA, AU, AC, etc.) as a
percentage for each sequence in an RNAStringSet.
}
