% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{get_sliding_windows}
\alias{get_sliding_windows}
\title{Get Sliding Windows of Upstream Regions}
\usage{
get_sliding_windows(
  upstr_ranges,
  window_width = 10,
  width_upstream = 100,
  overlap = 5
)
}
\arguments{
\item{upstr_ranges}{A GRanges object containing upstream regions.}

\item{window_width}{Integer width of each window (default is 10).}

\item{width_upstream}{Total width of upstream region (default is 100); must match flankupsteam.}

\item{overlap}{Number of overlapping nucleotides between adjacent windows (default is 5).}
}
\value{
A matrix of features created from overlapping windows.
}
\description{
Divides upstream sequences (as a GRanges object) into overlapping windows and
calculates the percentage of each nucleotide per window.
}
