% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{get_tribp_features}
\alias{get_tribp_features}
\title{Get Tri-nucleotide Features}
\usage{
get_tribp_features(stringSet)
}
\arguments{
\item{stringSet}{An RNAStringSet object.}
}
\value{
A matrix with percentages for each tri-nucleotide.
}
\description{
Calculates the frequency of tri-nucleotides (triplets, e.g. AAA, AAU, AAC, etc.) as a
percentage for each sequence in an RNAStringSet.
}
