% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_postDTU.R
\name{get_windows}
\alias{get_windows}
\title{Get Windows of Upstream Regions}
\usage{
get_windows(upstr_ranges, window_width = 10, width_upstream = 100)
}
\arguments{
\item{upstr_ranges}{A GRanges object containing upstream regions.}

\item{window_width}{Integer width of each window (default is 10).}

\item{width_upstream}{Total width of upstream region (default is 100); must match flankupsteam.}
}
\value{
A matrix with nucleotide percentages for each window (features named by window and nucleotide).
}
\description{
Divides the upstream sequences (provided as a GRanges object) into non-overlapping windows
and calculates the percentage of each nucleotide per window.
}
