% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLain-package.R
\docType{package}
\name{SPLain-package}
\alias{SPLain}
\alias{SPLain-package}
\title{SPLain: DTU}
\description{
Some description...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beamimc/SPLain}
  \item \url{https://Beamimc.github.io/SPLain}
  \item Report bugs at \url{https://github.com/beamimc/SPLain/issues}
}

}
\author{
\strong{Maintainer}: Beatriz Campillo Miñano \email{beacm@unc.edu}

Authors:
\itemize{
  \item Michael Love \email{michaelisaiahlove@gmail.com} (\href{https://orcid.org/0000-0001-8401-0545}{ORCID})
}

}
\keyword{internal}
